+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Minified add-floating-scrollbar.js.
+
+### Fixed
+- No scrolling after refreshing interactive report.
+
 ## 1.0.0 - 2017-02-16
 ### Added
 - Dynamic action plugin based on Ben Alman's [jQuery Floating Scrollbar](https://gist.github.com/cowboy/846423).
+
+[Unreleased]: https://github.com/foobarbaz-pl/apex-plugin-floating-scrollbar/compare/v1.0.0...HEAD